requirements.txt
streamlit>=1.26.0
openai>=1.3.5
pyyaml>=6.0
python-dotenv>=1.0.0
watchdog>=3.0.0